+++ src/Symfony/Bundle/FrameworkBundle/Command/SecretsDecryptToLocalCommand.php
@@ -38,6 +38,7 @@ final class SecretsDecryptToLocalCommand extends Command
     protected function configure(): void
     {
         $this
+            ->addOption('exit', null, InputOption::VALUE_NONE, 'Returns a non-zero exit code if any errors are encountered')
             ->addOption('force', 'f', InputOption::VALUE_NONE, 'Force overriding of secrets that already exist in the local vault')
             ->setHelp(<<<'EOF'
 The <info>%command.name%</info> command decrypts all secrets and copies them in the local vault.
@@ -47,6 +48,10 @@ The <info>%command.name%</info> command decrypts all secrets and copies them in 
 When the <info>--force</info> option is provided, secrets that already exist in the local vault are overridden.
 
     <info>%command.full_name% --force</info>
+
+When the <info>--exit</info> option is provided, the command will return a non-zero exit code if any errors are encountered.
+
+    <info>%command.full_name% --exit</info>
 EOF
             )
         ;
@@ -64,7 +69,7 @@ EOF
 
         $secrets = $this->vault->list(true);
 
-        $io->comment(sprintf('%d secret%s found in the vault.', \count($secrets), 1 !== \count($secrets) ? 's' : ''));
+        $io->comment(\sprintf('%d secret%s found in the vault.', \count($secrets), 1 !== \count($secrets) ? 's' : ''));
 
         $skipped = 0;
         if (!$input->getOption('force')) {
@@ -78,19 +83,25 @@ EOF
 
         if ($skipped > 0) {
             $io->warning([
-                sprintf('%d secret%s already overridden in the local vault and will be skipped.', $skipped, 1 !== $skipped ? 's are' : ' is'),
+                \sprintf('%d secret%s already overridden in the local vault and will be skipped.', $skipped, 1 !== $skipped ? 's are' : ' is'),
                 'Use the --force flag to override these.',
             ]);
         }
 
+        $hadErrors = false;
         foreach ($secrets as $k => $v) {
             if (null === $v) {
-                $io->error($this->vault->getLastMessage() ?? sprintf('Secret "%s" has been skipped as there was an error reading it.', $k));
+                $io->error($this->vault->getLastMessage() ?? \sprintf('Secret "%s" has been skipped as there was an error reading it.', $k));
+                $hadErrors = true;
                 continue;
             }
 
             $this->localVault->seal($k, $v);
             $io->note($this->localVault->getLastMessage());
+        }
+
+        if ($hadErrors && $input->getOption('exit'))  {
+            return 1;
         }
 
         return 0;