+++ src/Symfony/Component/Validator/Constraints/BicValidator.php
@@ -100,6 +100,9 @@ class BicValidator extends ConstraintValidator
         }
 
         $bicCountryCode = substr($canonicalize, 4, 2);
+        if (Bic::VALIDATION_MODE_CASE_INSENSITIVE === $constraint->mode) {
+            $bicCountryCode = strtoupper($bicCountryCode);
+        }
         if (!isset(self::BIC_COUNTRY_TO_IBAN_COUNTRY_MAP[$bicCountryCode]) && !Countries::exists($bicCountryCode)) {
             $this->context->buildViolation($constraint->message)
                 ->setParameter('{{ value }}', $this->formatValue($value))
@@ -109,8 +112,8 @@ class BicValidator extends ConstraintValidator
             return;
         }
 
-        // should contain uppercase characters only
-        if (strtoupper($canonicalize) !== $canonicalize) {
+        // should contain uppercase characters only in strict mode
+        if (Bic::VALIDATION_MODE_STRICT === $constraint->mode && strtoupper($canonicalize) !== $canonicalize) {
             $this->context->buildViolation($constraint->message)
                 ->setParameter('{{ value }}', $this->formatValue($value))
                 ->setCode(Bic::INVALID_CASE_ERROR)
@@ -126,7 +129,7 @@ class BicValidator extends ConstraintValidator
             try {
                 $iban = $this->getPropertyAccessor()->getValue($object, $path);
             } catch (NoSuchPropertyException $e) {
-                throw new ConstraintDefinitionException(sprintf('Invalid property path "%s" provided to "%s" constraint: ', $path, get_debug_type($constraint)).$e->getMessage(), 0, $e);
+                throw new ConstraintDefinitionException(\sprintf('Invalid property path "%s" provided to "%s" constraint: ', $path, get_debug_type($constraint)).$e->getMessage(), 0, $e);
             }
         }
         if (!$iban) {