+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+7.2
+---
+
+ * `WrappedExceptionsInterface` now extends PHP's `Throwable` interface
+ * Add `#[AsMessage]` attribute with `$transport` parameter for message routing
+ * Add `--format` option to the `messenger:stats` command
+
 7.1
 ---